% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by_period.R
\name{group_by_period}
\alias{group_by_period}
\title{Aggregate a numeric column over a specified time grouping.}
\usage{
group_by_period(.data, .date_col, .summarise_col, unit, agg_fun = sum,
  na.rm = FALSE)
}
\arguments{
\item{.data}{A \code{data.frame} to apply \code{group_by_period} to}

\item{.date_col}{A date column that will be used for grouping the date.
This should be of type \code{date} or \code{datetime}}

\item{.summarise_col}{The column which will be aggregated}

\item{unit}{A character string specifying a time unit or a multiple of
a unit to be rounded to. See \code{?lubridate::floor_date} for more details}

\item{agg_fun}{A function used to aggregate the numeric column}

\item{na.rm}{\code{NA} handling - this is passed to the \code{na.rm} argument of the
aggregating function}
}
\description{
\code{group_by_period} takes a date column and a numeric column and aggregates
the numeric column over specified date period. This is a common pattern in
data analysis in business where we want to convert transactional data to
daily, weekly or monthly data.
}
\examples{

library(nycflights13)
library(tidyr)
library(dplyr)

data("flights")

flights_w_date <- flights \%>\%
  unite(flight_date, year, month, day, sep = "-") \%>\%
  mutate(flight_date = ymd(flight_date))

flights_w_date \%>\%
  group_by_period(flight_date, distance, "month")

}
