% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_top_n.R
\name{keep_top_n}
\alias{keep_top_n}
\title{Filter a dataframe to keep the top n of a given column.}
\usage{
keep_top_n(.data, .ranking_col, n = 1)
}
\arguments{
\item{.data}{A \code{data.frame} to apply the filter method too}

\item{.ranking_col}{The column used to filter the data}

\item{n}{The number of rows to keep}
}
\description{
\code{keep_top_n} takes a data frame and filters it to only return the rows
  coinciding with the top n of a given column. This enables the user to quickly
  extract data based on the \emph{first} or \emph{largest} of something.
}
\details{
The ranking mechanism uses the \code{dplyr::row_number}. This means if there
  are any ties when ranking the data, the data corresponding to the first
  row positionally is kept.
}
\examples{

library(dplyr)

data("starwars")

starwars \%>\%
  group_by(gender) \%>\%
  keep_top_n(height)

starwars \%>\%
  group_by(gender) \%>\%
  keep_top_n(desc(height), n = 3)


}
