% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_multi.R
\name{spread_multi}
\alias{spread_multi}
\title{Spread multiple columns simultaneously.}
\usage{
spread_multi(.data, .key, ..., sep = "_")
}
\arguments{
\item{.data}{a data.frame to be spread}

\item{.key}{the key column which will be spread across the columns of the
resulting data frame}

\item{...}{the value columns to be spread. All column naming conventions that
can be used with \code{dplyr::select} should behave well with
\code{spread_multi}}

\item{sep}{the seperator to create column names of the form
`<key_name><sep><key_value>`}
}
\description{
\code{spread_multi} works simililarly to \code{tidyr::spread} but allows
for multiple value columns to be passed to the ellipsis (\code{...}) argument
which are spread simultaneously
}
\examples{

df <- data_frame(
  month = rep(1:3,2),
  student = rep(c("Amy", "Bob"), each=3),
  x = c(9, 7, 6, 8, 6, 9),
  y = c(6, 7, 8, 5, 6, 7),
  z = c(10, 5, 3, 2, 1, -1)
)

df \%>\%
  spread_multi(student, x:z)


}
